requirements.txt
tensorflow
numpy
matplotlib
pillow
gradio